--- a/docs/_includes/top.html
+++ b/docs/_includes/top.html
@@ -6,7 +6,6 @@
   <meta name="generator" content="Jekyll v{{ jekyll.version }}">
   {% feed_meta %}
   <link rel="alternate" type="application/atom+xml" title="Recent commits to Jekyll’s master branch" href="{{ site.repository }}/commits/master.atom">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/docsearch.js@2/dist/cdn/docsearch.min.css" />
   <link rel="preload" href="/fonts/lato-v14-latin-300.woff2" as="font" type="font/woff2" crossorigin />
   <link rel="preload" href="/fonts/lato-v14-latin-700.woff2" as="font" type="font/woff2" crossorigin />
   <link rel="stylesheet" href="/css/screen.css">