--- a/docs/css/screen.scss
+++ b/docs/css/screen.scss
@@ -6,5 +6,6 @@
 @import "gridism";
 @import "pygments";
 @import "font-awesome";
-@import "style";
+@import "fonts";
 @import "docsearch";
+@import "style";