This is a 🔦 documentation change.

TL;DR: PageSpeed report went from 94 to 100

## Context

Current PageSpeed report: https://developers.google.com/speed/pagespeed/insights/?url=jekyllrb.com

![pagespeed-current](https://user-images.githubusercontent.com/103008/48783611-2063ac80-ece1-11e8-93f2-e3dbb1b30da1.png)

Speed Index: 2.3s
Time to Interactive: 3.2s
First Meaningul Paint; 2.3s 

## Changes

Make Jekyll website load a little bit faster by:

- self-hosting fonts with modern formats (woff, woff2)
- preloading important fonts earlier 
- prioritize readable text
- self-hosting docsearch CSS to avoid network latency
- preloading main CSS to avoid blocking rendering
- resize and optimize logos to have a lighter total

New PageSpeed report: https://developers.google.com/speed/pagespeed/insights/?url=https%3A%2F%2Fdeploy-preview-7388--jekyllrb.netlify.com%2F

![pagespeed-7138](https://user-images.githubusercontent.com/103008/48783725-628cee00-ece1-11e8-8038-b73f81f712e6.png)




Speed Index: 1.3s
Time to Interactive: 2.4s
First Meaningul Paint; 1s