@DirtyF 2018-11-20
@jekyllbot: merge +docs