@DirtyF 2018-11-20
@Parker Thanks, next time we'll try to aim at a perfect score at Progressive Web App.